// src/app/Models/DemoStudent.ts
export class DemoStudent {
  _id: string = '';
  referenceId: string = '';
  fullName: string = '';
  fatherName: string = '';
  category: string = '';
  studentContact: string = '';
  class: string = ''; // Students Class
  medium: string = ''; // Students Medium
  state: string = ''; // Students state
  district: string = ''; // Students district
  address: string = ''; //Students address
  referenceMedium: string = ''; // Student Referenced From
  counsellorName: string = ''; // counseller name
  counsellorCabin: string = '';
  createdAt: Date = new Date();
}